// src/public/model/enumeration/dnd5/ClassesEnum.ts
export enum ClassesEnum {
    BARBARIAN = 'Barbarian',
    BARD = 'Bard',
    CLERIC = 'Cleric',
    DRUID = 'Druid',
    FIGHTER = 'Fighter',
    MONK = 'Monk',
    PALADIN = 'Paladin',
    RANGER = 'Ranger',
    ROGUE = 'Rogue',
    SORCERER = 'Sorcerer',
    WARLOCK = 'Warlock',
    WIZARD = 'Wizard',
    ARTIFICER = 'Artificer',
    NONE = '',
}